flask_app/app/requirements.txt
absl-py==0.7.1
aniso8601==6.0.0
astor==0.7.1
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
decorator==4.4.0
Flask==1.0.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gast==0.2.2
grpcio==1.20.1
gunicorn
h5py==2.9.0
idna==2.8
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter-core==4.4.0
Markdown==3.1
MarkupSafe==1.1.1
mock==3.0.3
nbformat==4.4.0
nltk==3.4.5
numpy==1.16.3
pandas==0.24.2
plotly==3.8.1
protobuf==3.7.1
pyrsistent==0.15.1
python-dateutil==2.8.0
python-slugify==3.0.0
pytz==2019.1
PyYAML==5.1
requests==2.21.0
retrying==1.3.3
scipy==1.2.1
six==1.12.0
scikit-learn==0.20.2
SQLAlchemy==1.3.3
termcolor==1.1.0
text-unidecode==1.2
traitlets==4.3.2
urllib3==1.24.3
Werkzeug==0.15.3
WTForms==2.2.1
flask_dropzone
flask_uploads
google-cloud-storage==1.37.1
absl-py

